detailed-event responzivitás+window.confirm+apróbb vizuális módosítások

+++ src/app/events-admin-list/events-admin-list.component.ts
@@ -11,10 +11,10 @@ import { BaseService } from '../base.service';
 export class EventsAdminListComponent {
   // albums:any=[]
   // szortirozottAdatok = new Subject()
-  addAddColumn:any
-  addEditColumn:any
-  addDeleteColumn:any
-  errMessage:any
+  addAddColumn: any
+  addEditColumn: any
+  addDeleteColumn: any
+  errMessage: any
 
   name = ""
   startDate = ""
@@ -37,30 +37,30 @@ export class EventsAdminListComponent {
   erModEndTime: any
 
   //A táblázat megjelenítéséhez
-  newEvent ={image: '', name: '', description: '', startDate: '', endDate: '', startTime: '', endTime: '', locationName: '', locationcountry: '', address: '', gpx: '', weblink: ''}
-  events:any
+  newEvent = { image: '', name: '', description: '', startDate: '', endDate: '', startTime: '', endTime: '', locationName: '', locationcountry: '', address: '', gpx: '', weblink: '' }
+  events: any
 
   // adatok módosításához
   editModeId: number | null = null
 
   //módosításkor fellépő hibaüzenetek elmentése
-  errModfyMsg : any
+  errModfyMsg: any
   //új esemény felvételekor fellépő hibaüzenetek tárolása
-  errNewEventMsg:any
+  errNewEventMsg: any
 
-  constructor(public base:BaseService,
-      // private config:ConfigService,
-      private http:HttpClient,
-      public localStorage:LocalStorageService) {
+  constructor(public base: BaseService,
+    // private config:ConfigService,
+    private http: HttpClient,
+    public localStorage: LocalStorageService) {
 
     this.base.getAll()
     this.getDataFromApi()
     this.base.downloadAll()
   }
 
-  getDataFromApi(){
+  getDataFromApi() {
     this.base.getAll().subscribe(
-      (res:any) => {
+      (res: any) => {
         this.events = res.data
       }
     )
@@ -74,14 +74,14 @@ export class EventsAdminListComponent {
   cancelEdit() {
     this.editModeId = null
     this.base.downloadAll()
-  } 
+  }
 
-  updateData(data:any){
+  updateData(data: any) {
     this.base.updateDataWeb(data).subscribe(
       {
-        next:(res:any)=>{
-          if(res.success == false){
-            console.log("hibaüzenetek: ",res.error)
+        next: (res: any) => {
+          if (res.success == false) {
+            console.log("hibaüzenetek: ", res.error)
             this.errModfyMsg = res.error
             this.erModName = ""
             this.erModName = res.error["name"]
@@ -89,18 +89,18 @@ export class EventsAdminListComponent {
             this.erModEndDate = res.error["endDate"]
             this.erModStartTime = res.error["startTime"]
             this.erModEndTime = res.error["endTime"]
-            this.base.show( res.message ||"Hiba történt!", "danger")
+            this.base.show(res.message || "Hiba történt!", "danger")
           }
           else {
-          console.log("Sikeres módosítás", res)
-          this.eventModifySuccess = true
-          this.base.show(res.message || "Sikeres módosítás!", "success")
-          this.editModeId = null
-          this.base.downloadAll()
+            console.log("Sikeres módosítás", res)
+            this.eventModifySuccess = true
+            this.base.show(res.message || "Sikeres módosítás!", "success")
+            this.editModeId = null
+            this.base.downloadAll()
           }
         },
-        error:(error:any)=>{
-          console.log("Valami hiba: ",error)
+        error: (error: any) => {
+          console.log("Valami hiba: ", error)
           this.base.show("Hálózati hiba vagy szerverhiba történt!", "danger")
         }
       }
@@ -108,31 +108,35 @@ export class EventsAdminListComponent {
     console.log("data" + data);
   }
 
-  deleteData(data:any){
-    this.base.deleteDataWeb(data).subscribe(
-      {
-        next:(res:any)=>{
-          //console.log("sikeres törlés: ",res)
-          this.base.downloadAll()
-          this.base.show(res.message || "Sikeres törlés!", "success")
-        },
-        error:(error:any)=>{
-          //console.log("Valami hiba: ",error)
-          this.base.show("Hálózati hiba vagy szerverhiba történt!", "danger")
+  deleteData(data: any) {
+    const confirmDelete = window.confirm('Biztosan törlöd ezt az eseményt?')
+
+    if (confirmDelete) {
+      this.base.deleteDataWeb(data).subscribe(
+        {
+          next: (res: any) => {
+            //console.log("sikeres törlés: ",res)
+            this.base.downloadAll()
+            this.base.show(res.message || "Sikeres törlés!", "success")
+          },
+          error: (error: any) => {
+            //console.log("Valami hiba: ",error)
+            this.base.show("Hálózati hiba vagy szerverhiba történt!", "danger")
+          }
         }
-      }
-    )
-    //console.log("data" + data);
+      )
+    }
+
 
   }
 
-  newData(){
+  newData() {
     this.base.newDataWeb(this.newEvent).subscribe(
       {
-        next:(res:any)=>{
+        next: (res: any) => {
           //console.log("új esemény felvétele: ",res)
-          if(res.success == false){
-            console.log("hibaüzenetek: ",res.error)
+          if (res.success == false) {
+            console.log("hibaüzenetek: ", res.error)
             this.errNewEventMsg = res.error
             this.erName = ""
             this.erName = res.error["name"]
@@ -140,10 +144,10 @@ export class EventsAdminListComponent {
             this.erEndDate = res.error["endDate"]
             this.erStartTime = res.error["startTime"]
             this.erEndTime = res.error["endTime"]
-            this.base.show( "Hiba történt!", "danger")
+            this.base.show("Hiba történt!", "danger")
           }
           //ahoz hogy az oldal újrafrissüljön.
-          else{
+          else {
             //console.log("Sikeres új esemény felvétel: ",res)
             this.base.show(res.message || "Sikeres rögzítés!", "success")
             this.newEventSuccess = true
@@ -151,13 +155,13 @@ export class EventsAdminListComponent {
           }
 
         },
-        error:(error:any)=>{
+        error: (error: any) => {
           // console.log("Valami hiba történt az új esemény felvétele során: ",error)
           this.base.show("Hálózati hiba vagy szerverhiba történt!", "danger")
         }
       }
     )
-    this.newEvent = {image: '', name: '', description: '', startDate: '', endDate: '', startTime: '', endTime: '', locationName: '', locationcountry: '', address: '', gpx: '', weblink: ''}
+    this.newEvent = { image: '', name: '', description: '', startDate: '', endDate: '', startTime: '', endTime: '', locationName: '', locationcountry: '', address: '', gpx: '', weblink: '' }
   }
 
 }
